refactor: all actions in actions.rs
