+++ src/actions/write.rs
@@ -1,4 +1,6 @@
-use crate::{result_ok, CommandStatus, Hrtor};
+use crate::{CommandStatus, Hrtor};
+
+use super::result_ok;
 
 impl Hrtor {
     pub(crate) fn write(&self) -> CommandStatus {
